[REFACTOR] - Moved controller classes to view
