+++ emoji-library/src/main/kotlin/br/com/instachat/emojilibrary/view/fragments/Nature.kt
@@ -1,4 +1,4 @@
-package br.com.instachat.emojilibrary.controller.pages
+package br.com.instachat.emojilibrary.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,21 +6,20 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.GridView
 import br.com.instachat.emojilibrary.R
-import br.com.instachat.emojilibrary.adapter.EmojiAdapter
-import br.com.instachat.emojilibrary.controller.FragmentEmoji
+import br.com.instachat.emojilibrary.view.adapter.EmojiAdapter
 import br.com.instachat.emojilibrary.model.Emoji
-import br.com.instachat.emojilibrary.model.People
+import br.com.instachat.emojilibrary.model.Nature
 import br.com.instachat.emojilibrary.util.Constants
-import java.util.*
 
-class People : FragmentEmoji() {
+class Nature : FragmentEmoji() {
+
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.frag_emoji_people, container, false)
+        return inflater.inflate(R.layout.frag_emoji_nature, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        var data: Array<Emoji> = People.DATA
+        var data: Array<Emoji> = Nature.DATA
         var useSystemDefault = false
         if (arguments != null) {
             data = arguments!!.getParcelableArray(Constants.EMOJI_KEY)?.map { it as Emoji }!!.toTypedArray()
@@ -28,7 +27,7 @@ class People : FragmentEmoji() {
         }
 
         val gridView = view.findViewById<GridView>(R.id.Emoji_GridView)
-        gridView.adapter = EmojiAdapter(view.context, Arrays.asList(*data), useSystemDefault)
+        gridView.adapter = EmojiAdapter(view.context, mutableListOf(*data), useSystemDefault)
         gridView.onItemClickListener = this
     }
 }
