+++ emoji-library/src/main/kotlin/br/com/instachat/emojilibrary/view/fragments/Recents.kt
@@ -1,4 +1,4 @@
-package br.com.instachat.emojilibrary.controller.pages
+package br.com.instachat.emojilibrary.view.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,27 +6,38 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.GridView
 import br.com.instachat.emojilibrary.R
-import br.com.instachat.emojilibrary.adapter.EmojiAdapter
-import br.com.instachat.emojilibrary.controller.FragmentEmoji
+import br.com.instachat.emojilibrary.view.adapter.EmojiAdapter
+import br.com.instachat.emojilibrary.view.fragments.FragmentEmoji
 import br.com.instachat.emojilibrary.model.Emoji
-import br.com.instachat.emojilibrary.model.Symbols
 import br.com.instachat.emojilibrary.util.Constants
+import br.com.instachat.emojilibrary.view.listeners.EmojiClickListener
 
-class Symbols : FragmentEmoji() {
+class Recents : FragmentEmoji(), EmojiClickListener {
+
+    private lateinit var adapter: EmojiAdapter
+
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.frag_emoji_symbols, container, false)
+        return inflater.inflate(R.layout.frag_emoji_recents, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        var data: Array<Emoji> = Symbols.DATA
         var useSystemDefault = false
         if (arguments != null) {
-            data = arguments!!.getParcelableArray(Constants.EMOJI_KEY)?.map { it as Emoji }!!.toTypedArray()
             useSystemDefault = arguments!!.getBoolean(Constants.USE_SYSTEM_DEFAULT_KEY)
         }
         val gridView = view.findViewById<GridView>(R.id.Emoji_GridView)
-        gridView.adapter = EmojiAdapter(view.context, mutableListOf(*data), useSystemDefault)
+        adapter = EmojiAdapter(view.context, listRecents(), useSystemDefault)
+        gridView.adapter = adapter
         gridView.onItemClickListener = this
+    }
+
+    override fun onResume() {
+        super.onResume()
+        adapter.updateItems(listRecents())
+    }
+
+    override fun emojiCliked(emojicon: Emoji?) {
+        adapter.notifyDataSetChanged()
     }
 }
