+++ emoji-library/src/main/kotlin/br/com/instachat/emojilibrary/view/listeners/EmojiClickListener.kt
@@ -2,6 +2,6 @@ package br.com.instachat.emojilibrary.view.listeners
 
 import br.com.instachat.emojilibrary.model.Emoji
 
-interface OnEmojiClickListener {
+interface EmojiClickListener {
     fun emojiCliked(emojicon: Emoji?)
 }
