--- emoji-library/src/main/kotlin/br/com/instachat/emojilibrary/view/listeners/WhatsAppPanelEventListener.kt
@@ -1,5 +1,0 @@
-package br.com.instachat.emojilibrary.view.listeners
-
-interface WhatsAppPanelEventListener {
-    fun onSendClicked()
-}
